architecture: Use Jekyll to build site.
My belief is that something should not be made harder to do just for
the sake of being seen as 'hard core', and in a similar fashion, one
must also not be expedient in their choices.

Since the need currently is only for a static website to display my
learnings, thoughts and opinions on certain topics, I will not go as
far as to build everything by hand, write down every piece of code
which has already been written as though that will prove
something. Rewriting the wheel is not my objective here. Building
something fresh for this blog is definitely on my mind, and I
probably will do a series for it, documenting my learning as I go.
But for now, a static site generator like Jekyll has to suffice.